fix: unknown config keys when initializing logger in system probe
